fix: keep name and description when switching create-reference mode
EmptyReference and ImportReference were two independent forms, so
switching between them lost whatever had been typed. Merge them into
one CreateReferenceForm with a single react-hook-form instance shared
across modes, and swap the empty/import tabs for a ToggleGroup mode
switch since these were never really independent content panes -
just two variants of the same form.
