+++ apps/web/src/references/components/CreateReference.tsx
@@ -1,8 +1,8 @@
 import { Dialog, DialogContent, DialogTitle } from "@base/Dialog";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@base/Tabs";
+import ToggleGroup from "@base/ToggleGroup";
+import ToggleGroupItem from "@base/ToggleGroupItem";
 import { useState } from "react";
-import EmptyReference from "./EmptyReference";
-import ImportReference from "./ImportReference";
+import { CreateReferenceForm } from "./CreateReferenceForm";
 
 type CreateReferenceProps = {
 	open: boolean;
@@ -13,47 +13,33 @@ type CreateReferenceProps = {
  * The create reference view with options to create an empty reference or import a reference
  */
 export function CreateReference({ open, onOpenChange }: CreateReferenceProps) {
-	const [tab, setTab] = useState<"empty" | "import">("empty");
+	const [mode, setMode] = useState<"empty" | "import">("empty");
 
 	function handleOpenChange(open: boolean) {
 		onOpenChange(open);
 		if (!open) {
-			setTab("empty");
+			setMode("empty");
 		}
 	}
 
 	function handleSuccess() {
 		onOpenChange(false);
-		setTab("empty");
+		setMode("empty");
 	}
 
 	return (
 		<Dialog open={open} onOpenChange={handleOpenChange}>
 			<DialogContent size="lg">
 				<DialogTitle>Create Reference</DialogTitle>
-				<Tabs
-					value={tab}
-					onValueChange={(value) => setTab(value as "empty" | "import")}
+				<ToggleGroup
+					className="mb-4"
+					value={mode}
+					onValueChange={(value) => setMode(value as "empty" | "import")}
 				>
-					<TabsList>
-						<TabsTrigger value="empty">Empty</TabsTrigger>
-						<TabsTrigger value="import">Import</TabsTrigger>
-					</TabsList>
-					<TabsContent
-						value="empty"
-						forceMount
-						className="data-[state=inactive]:hidden"
-					>
-						<EmptyReference onSuccess={handleSuccess} />
-					</TabsContent>
-					<TabsContent
-						value="import"
-						forceMount
-						className="data-[state=inactive]:hidden"
-					>
-						<ImportReference onSuccess={handleSuccess} />
-					</TabsContent>
-				</Tabs>
+					<ToggleGroupItem value="empty">Empty</ToggleGroupItem>
+					<ToggleGroupItem value="import">Import</ToggleGroupItem>
+				</ToggleGroup>
+				<CreateReferenceForm mode={mode} onSuccess={handleSuccess} />
 			</DialogContent>
 		</Dialog>
 	);
