+++ apps/web/src/references/components/CreateReferenceForm.tsx
@@ -1,0 +1,179 @@
+import Alert from "@base/Alert";
+import { DialogFooter } from "@base/Dialog";
+import InputError from "@base/InputError";
+import InputGroup from "@base/InputGroup";
+import InputLabel from "@base/InputLabel";
+import InputSimple from "@base/InputSimple";
+import ProgressBarAffixed from "@base/ProgressBarAffixed";
+import SaveButton from "@base/SaveButton";
+import TextArea from "@base/TextArea";
+import { useNavigate } from "@tanstack/react-router";
+import { useId } from "react";
+import { Controller, useForm } from "react-hook-form";
+import { UploadBar } from "@/uploads/components/UploadBar";
+import {
+	useCreateReference,
+	useImportReference,
+	useUploadReference,
+} from "../queries";
+
+type FormValues = {
+	name: string;
+	description: string;
+	organism: string;
+	upload: string;
+};
+
+type CreateReferenceFormProps = {
+	mode: "empty" | "import";
+	onSuccess: () => void;
+};
+
+/**
+ * A form for creating a reference from scratch or by importing a file, sharing
+ * name and description between the two modes
+ */
+export function CreateReferenceForm({
+	mode,
+	onSuccess,
+}: CreateReferenceFormProps) {
+	const navigate = useNavigate();
+	const createMutation = useCreateReference();
+	const importMutation = useImportReference();
+	const { uploadMutation, fileName, uploadId, progress } = useUploadReference();
+
+	const nameId = useId();
+	const organismId = useId();
+	const descriptionId = useId();
+
+	const {
+		control,
+		formState: { errors },
+		handleSubmit,
+		register,
+	} = useForm<FormValues>({
+		defaultValues: {
+			name: "",
+			description: "",
+			organism: "",
+			upload: "",
+		},
+	});
+
+	function handleDrop(acceptedFiles: File[]) {
+		const file = acceptedFiles[0];
+		if (file === undefined) {
+			return;
+		}
+
+		uploadMutation.mutate(file);
+	}
+
+	function onSubmit(values: FormValues) {
+		if (mode === "import") {
+			if (uploadId === null) {
+				return;
+			}
+
+			importMutation.mutate(
+				{
+					name: values.name,
+					description: values.description,
+					importFrom: uploadId,
+				},
+				{
+					onSuccess: () => {
+						navigate({ to: "/refs", replace: true });
+						onSuccess();
+					},
+				},
+			);
+
+			return;
+		}
+
+		createMutation.mutate(
+			{
+				name: values.name,
+				description: values.description,
+				organism: values.organism,
+			},
+			{
+				onSuccess: () => {
+					navigate({ to: "/refs" });
+					onSuccess();
+				},
+			},
+		);
+	}
+
+	const uploadBarMessage =
+		fileName || (progress === 0 ? "Drag file here to upload" : "Uploading...");
+
+	return (
+		<form onSubmit={handleSubmit(onSubmit)}>
+			{mode === "import" && (
+				<>
+					<Alert>
+						<strong>
+							Create a reference from a file previously exported from another
+							Virtool reference.
+						</strong>
+					</Alert>
+
+					<Controller
+						control={control}
+						name="upload"
+						rules={{ required: "A reference file must be uploaded" }}
+						render={({ field: { onChange } }) => (
+							<div className="mb-4">
+								<ProgressBarAffixed color="green" now={progress} />
+								<UploadBar
+									message={uploadBarMessage}
+									onDrop={(acceptedFiles) => {
+										handleDrop(acceptedFiles);
+										const file = acceptedFiles[0];
+										if (file) {
+											onChange(file.name);
+										}
+									}}
+									multiple={false}
+								/>
+
+								<InputError>{errors.upload?.message}</InputError>
+							</div>
+						)}
+					/>
+				</>
+			)}
+
+			<InputGroup className="pb-0">
+				<InputLabel htmlFor={nameId}>Name</InputLabel>
+				<InputSimple
+					id={nameId}
+					{...register("name", { required: "Required Field" })}
+				/>
+				<InputError>{errors.name?.message}</InputError>
+			</InputGroup>
+
+			{mode === "empty" && (
+				<InputGroup>
+					<InputLabel htmlFor={organismId}>Organism</InputLabel>
+					<InputSimple id={organismId} {...register("organism")} />
+				</InputGroup>
+			)}
+
+			<InputGroup>
+				<InputLabel htmlFor={descriptionId}>Description</InputLabel>
+				<TextArea id={descriptionId} {...register("description")} />
+			</InputGroup>
+
+			<DialogFooter>
+				<SaveButton
+					altText={mode === "import" ? "Import" : "Save"}
+					disabled={mode === "import" && progress > 0 && uploadId === null}
+				/>
+			</DialogFooter>
+		</form>
+	);
+}
