--- apps/web/src/references/components/ImportReference.tsx
@@ -1,127 +1,0 @@
-import Alert from "@base/Alert";
-import { DialogFooter } from "@base/Dialog";
-import InputError from "@base/InputError";
-import InputGroup from "@base/InputGroup";
-import InputLabel from "@base/InputLabel";
-import InputSimple from "@base/InputSimple";
-import ProgressBarAffixed from "@base/ProgressBarAffixed";
-import SaveButton from "@base/SaveButton";
-import { useNavigate } from "@tanstack/react-router";
-import { useId } from "react";
-import { Controller, useForm } from "react-hook-form";
-import { UploadBar } from "@/uploads/components/UploadBar";
-import { useImportReference, useUploadReference } from "../queries";
-
-type ImportReferenceProps = {
-	onSuccess?: () => void;
-};
-
-export default function ImportReference({ onSuccess }: ImportReferenceProps) {
-	const navigate = useNavigate();
-	const nameId = useId();
-	const descriptionId = useId();
-
-	const importMutation = useImportReference();
-	const { uploadMutation, fileName, uploadId, progress } = useUploadReference();
-
-	const {
-		control,
-		formState: { errors },
-		handleSubmit,
-		register,
-	} = useForm({
-		defaultValues: {
-			name: "",
-			description: "",
-			upload: "",
-		},
-	});
-
-	function handleDrop(acceptedFiles: File[]) {
-		const file = acceptedFiles[0];
-		if (file === undefined) {
-			return;
-		}
-
-		uploadMutation.mutate(file);
-	}
-
-	const uploadBarMessage =
-		fileName || (progress === 0 ? "Drag file here to upload" : "Uploading...");
-
-	return (
-		<>
-			<Alert>
-				<strong>
-					Create a reference from a file previously exported from another
-					Virtool reference.
-				</strong>
-			</Alert>
-
-			<Controller
-				control={control}
-				name="upload"
-				rules={{ required: true }}
-				render={({ field: { onChange } }) => (
-					<div className="mb-4">
-						<ProgressBarAffixed color="green" now={progress} />
-						<UploadBar
-							message={uploadBarMessage}
-							onDrop={(acceptedFiles) => {
-								handleDrop(acceptedFiles);
-								const file = acceptedFiles[0];
-								if (file) {
-									onChange(file.name);
-								}
-							}}
-							multiple={false}
-						/>
-
-						<InputError>
-							{errors.upload?.type === "required" &&
-								"A reference file must be uploaded"}
-						</InputError>
-					</div>
-				)}
-			/>
-
-			<form
-				onSubmit={handleSubmit((values) => {
-					if (uploadId === null) {
-						return;
-					}
-					importMutation.mutate(
-						{ ...values, importFrom: uploadId },
-						{
-							onSuccess: () => {
-								navigate({ to: "/refs", replace: true });
-								onSuccess?.();
-							},
-						},
-					);
-				})}
-			>
-				<InputGroup>
-					<InputLabel htmlFor={nameId}>Name</InputLabel>
-					<InputSimple id={nameId} {...register("name", { required: true })} />
-					<InputError>{errors.name?.type && "A name is required."}</InputError>
-				</InputGroup>
-				<InputGroup>
-					<InputLabel htmlFor={descriptionId}>Description</InputLabel>
-					<InputSimple
-						as="textarea"
-						id={descriptionId}
-						{...register("description")}
-					/>
-				</InputGroup>
-
-				<DialogFooter>
-					<SaveButton
-						disabled={progress > 0 && uploadId === null}
-						altText="Import"
-					/>
-				</DialogFooter>
-			</form>
-		</>
-	);
-}
